+++ component/processor.go
@@ -69,35 +69,29 @@ type ProcessorFactory interface {
 	// tests of any implementation of the Factory interface.
 	CreateDefaultConfig() config.Processor
 
-	// CreateTracesProcessor creates a trace processor based on this config.
+	// CreateTracesProcessor creates a TracesProcessor based on this config.
 	// If the processor type does not support tracing or if the config is not valid,
 	// an error will be returned instead.
-	CreateTracesProcessor(
-		ctx context.Context,
-		set ProcessorCreateSettings,
-		cfg config.Processor,
-		nextConsumer consumer.Traces,
-	) (TracesProcessor, error)
+	CreateTracesProcessor(ctx context.Context, set ProcessorCreateSettings, cfg config.Processor, nextConsumer consumer.Traces) (TracesProcessor, error)
 
-	// CreateMetricsProcessor creates a metrics processor based on this config.
+	// TracesProcessorStability gets the stability level of the TracesProcessor.
+	TracesProcessorStability() StabilityLevel
+
+	// CreateMetricsProcessor creates a MetricsProcessor based on this config.
 	// If the processor type does not support metrics or if the config is not valid,
 	// an error will be returned instead.
-	CreateMetricsProcessor(
-		ctx context.Context,
-		set ProcessorCreateSettings,
-		cfg config.Processor,
-		nextConsumer consumer.Metrics,
-	) (MetricsProcessor, error)
+	CreateMetricsProcessor(ctx context.Context, set ProcessorCreateSettings, cfg config.Processor, nextConsumer consumer.Metrics) (MetricsProcessor, error)
 
-	// CreateLogsProcessor creates a processor based on the config.
+	// MetricsProcessorStability gets the stability level of the MetricsProcessor.
+	MetricsProcessorStability() StabilityLevel
+
+	// CreateLogsProcessor creates a LogsProcessor based on the config.
 	// If the processor type does not support logs or if the config is not valid,
 	// an error will be returned instead.
-	CreateLogsProcessor(
-		ctx context.Context,
-		set ProcessorCreateSettings,
-		cfg config.Processor,
-		nextConsumer consumer.Logs,
-	) (LogsProcessor, error)
+	CreateLogsProcessor(ctx context.Context, set ProcessorCreateSettings, cfg config.Processor, nextConsumer consumer.Logs) (LogsProcessor, error)
+
+	// LogsProcessorStability gets the stability level of the LogsProcessor.
+	LogsProcessorStability() StabilityLevel
 }
 
 // ProcessorCreateDefaultConfigFunc is the equivalent of ProcessorFactory.CreateDefaultConfig().
@@ -178,6 +172,18 @@ type processorFactory struct {
 	CreateLogsProcessorFunc
 }
 
+func (p processorFactory) TracesProcessorStability() StabilityLevel {
+	return p.getStabilityLevel(config.TracesDataType)
+}
+
+func (p processorFactory) MetricsProcessorStability() StabilityLevel {
+	return p.getStabilityLevel(config.MetricsDataType)
+}
+
+func (p processorFactory) LogsProcessorStability() StabilityLevel {
+	return p.getStabilityLevel(config.LogsDataType)
+}
+
 // WithTracesProcessor overrides the default "error not supported" implementation for CreateTracesProcessor and the default "undefined" stability level.
 func WithTracesProcessor(createTracesProcessor CreateTracesProcessorFunc, sl StabilityLevel) ProcessorFactoryOption {
 	return processorFactoryOptionFunc(func(o *processorFactory) {
@@ -185,9 +191,6 @@ func WithTracesProcessor(createTracesProcessor CreateTracesProcessorFunc, sl Sta
 		o.CreateTracesProcessorFunc = createTracesProcessor
 	})
 }
-
-// Deprecated: [v0.57.0] Use WithTracesProcessor instead.
-var WithTracesProcessorAndStabilityLevel = WithTracesProcessor
 
 // WithMetricsProcessor overrides the default "error not supported" implementation for CreateMetricsProcessor and the default "undefined" stability level.
 func WithMetricsProcessor(createMetricsProcessor CreateMetricsProcessorFunc, sl StabilityLevel) ProcessorFactoryOption {
@@ -197,9 +200,6 @@ func WithMetricsProcessor(createMetricsProcessor CreateMetricsProcessorFunc, sl 
 	})
 }
 
-// Deprecated: [v0.57.0] Use WithMetricsProcessor instead.
-var WithMetricsProcessorAndStabilityLevel = WithMetricsProcessor
-
 // WithLogsProcessor overrides the default "error not supported" implementation for CreateLogsProcessor and the default "undefined" stability level.
 func WithLogsProcessor(createLogsProcessor CreateLogsProcessorFunc, sl StabilityLevel) ProcessorFactoryOption {
 	return processorFactoryOptionFunc(func(o *processorFactory) {
@@ -207,9 +207,6 @@ func WithLogsProcessor(createLogsProcessor CreateLogsProcessorFunc, sl Stability
 		o.CreateLogsProcessorFunc = createLogsProcessor
 	})
 }
-
-// Deprecated: [v0.57.0] Use WithLogsProcessor instead.
-var WithLogsProcessorAndStabilityLevel = WithLogsProcessor
 
 // NewProcessorFactory returns a ProcessorFactory.
 func NewProcessorFactory(cfgType config.Type, createDefaultConfig ProcessorCreateDefaultConfigFunc, options ...ProcessorFactoryOption) ProcessorFactory {