+++ CHANGELOG.md
@@ -4,6 +4,38 @@
 
 ### 🛑 Breaking changes 🛑
 
+- Remove deprecated component stability helpers (#5802):
+  - `component.WithTracesExporterAndStabilityLevel`
+  - `component.WithMetricsExporterAndStabilityLevel`
+  - `component.WithLogsExporterAndStabilityLevel`
+  - `component.WithTracesReceiverAndStabilityLevel`
+  - `component.WithMetricsReceiverAndStabilityLevel`
+  - `component.WithLogsReceiverAndStabilityLevel`
+  - `component.WithTracesProcessorAndStabilityLevel`
+  - `component.WithMetricsProcessorAndStabilityLevel`
+  - `component.WithLogsProcessorAndStabilityLevel`
+
+### 🚩 Deprecations 🚩
+
+- Deprecate the `component.Factory.StabilityLevel(config.DataType)` in favor of Stability per component (#5762):
+  - `component.ExporterFactory.TracesExporterStability`
+  - `component.ExporterFactory.MetricsExporterStability`
+  - `component.ExporterFactory.LogsExporterStability`
+  - `component.ProcessorFactory.TracesProcessorStability`
+  - `component.ProcessorFactory.MetricsProcessorStability`
+  - `component.ProcessorFactory.LogsProcessorStability`
+  - `component.ReceiverFactory.TracesReceiverStability`
+  - `component.ReceiverFactory.MetricsReceiverStability`
+  - `component.ReceiverFactory.LogsReceiverStability`
+
+### 💡 Enhancements 💡
+
+### 🧰 Bug fixes 🧰
+
+## v0.57.0 Beta
+
+### 🛑 Breaking changes 🛑
+
 - Remove deprecated funcs/types from service related to `Config` (#5755)
 - Change`confighttp.ToClient` to accept a `component.Host` (#5737)
 - Remove deprecated funcs from pdata related to mutable slices (#5754)
@@ -17,6 +49,7 @@
   - `component.WithTracesProcessor`
   - `component.WithMetricsProcessor`
   - `component.WithLogsProcessor`
+- Update minimum go version to 1.18 (#5795)
 
 ### 🚩 Deprecations 🚩